+++ lib/release_packer.dart
@@ -2,3 +2,4 @@
 library release_updater.packer;
 
 export 'src/release_packer.dart';
+export 'src/release_updater_platform.dart';