+++ lib/release_updater_io.dart
@@ -4,4 +4,6 @@ library release_updater.io;
 export 'release_updater.dart';
 export 'src/release_updater_io.dart';
 export 'src/release_updater_isolate.dart';
+export 'src/release_updater_platform.dart';
 export 'src/release_updater_storage_io.dart';
+export 'src/release_updater_utils_io.dart';