+++ lib/src/release_updater_utils_io.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+extension FileExtension on File {
+  bool get hasExecutablePermission => statSync().modeString().contains('x');
+}